Remove Rates from listing
Also, the SummaryCardComponent was inheriting from a component which is
no longer used, so I’ve moved all the logic to the index component.
