--- lib/engines/content_block_manager/app/components/content_block_manager/content_block/document/show/summary_list_component.html.erb
@@ -1,3 +1,0 @@
-<%= render "govuk_publishing_components/components/summary_list", {
-  items:,
-} %>
